some renames in the test folder
