+++ .travis.yml
@@ -7,6 +7,6 @@ install:
 before_script:
     - mkdir logs
 script:
-    - coverage run --source=kitty unit_tests/runner.py
+    - coverage run --source=kitty tests/runner.py
 after_success:
     - coveralls
