+++ tests/test_fuzzer_client.py
